webdriverio: move shadowFnFactory to scripts
